feat: add a layout for the feed/profiles order list and modal logic for orders
