+++ src/components/app/pages/feed-orders/feed-orders.module.css
@@ -1,10 +1,5 @@
-.orders {
-  display: flex;
-  gap: 60px;
-}
-
 .order_feed_wrapper {
-  max-width: 844px;
+  max-width: 608px;
   composes: mt-10 from global;
 }
 
@@ -70,4 +65,40 @@
   display: flex;
   align-items: center;
   gap: 8px;
+}
+
+.statistics_panel {
+  display: flex;
+  flex-direction: column;
+  gap: 60px;
+  width: 580px;
+  composes: mt-25 from global;
+}
+
+.statistics_table {
+  display: flex;
+  max-height: 206px;
+}
+
+.statistics_orders_list {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: wrap;
+  gap: 8px;
+  max-height: 184px;
+}
+
+.statistics_wrapper {
+  width: 272px;
+}
+
+.text_blue {
+  color: #00cccc;
+}
+
+.digits_glow {
+  text-shadow:
+    0 4px 32px rgba(51, 51, 255, 0.5),
+    0 0 8px rgba(51, 51, 255, 0.25),
+    0 0 16px rgba(51, 51, 255, 0.25);
 }
