+++ src/components/app/pages/orders/orders.tsx
@@ -4,20 +4,29 @@ import {
 } from "@ya.praktikum/react-developer-burger-ui-components";
 import styles from "./orders.module.css";
 import test from "../../../../images/illustration.png";
+import { useModal } from "../../../../hooks/useModal";
+import { Outlet } from "react-router-dom";
 
 export default function Orders() {
   const format_date = () => {
     const date_from_server = "2022-10-10T17:33:32.877Z";
     return <FormattedDate date={new Date(date_from_server)} />;
   };
+  const { openModal } = useModal();
 
   return (
     <>
+      <Outlet />
       <article className={styles.order_feed_wrapper}>
-        <p className="text text_type_main-large">Лента заказов</p>
         <ul className={styles.order_items_list}>
           {[...Array(6)].map((_, i) => (
-            <li key={i} className={styles.order_item_card}>
+            <li
+              key={i}
+              className={styles.order_item_card}
+              onClick={() => {
+                openModal({ _id: 1 }, "profile-order");
+              }}
+            >
               <div className={styles.order_item_header}>
                 <p className="text text_type_digits-default">#123456</p>
                 <p className="text text_type_main-default text_color_inactive">
@@ -48,52 +57,6 @@ export default function Orders() {
             </li>
           ))}
         </ul>
-      </article>
-      <article className={styles.statistics_panel}>
-        <section className={styles.statistics_table}>
-          <div className={styles.statistics_wrapper}>
-            <p className="text text_type_main-medium mb-6">Готовы:</p>
-            <ul className={styles.statistics_orders_list}>
-              {[...Array(10)].map((_, i) => (
-                <li key={i}>
-                  <p
-                    className={
-                      styles.text_blue + " " + "text text_type_digits-default"
-                    }
-                  >
-                    123456
-                  </p>
-                </li>
-              ))}
-            </ul>
-          </div>
-          <div className={styles.statistics_wrapper}>
-            <p className="text text_type_main-medium mb-6">В работе:</p>
-            <ul className={styles.statistics_orders_list}>
-              {[...Array(10)].map((_, i) => (
-                <li key={i}>
-                  <p className="text text_type_digits-default">123456</p>
-                </li>
-              ))}
-            </ul>
-          </div>
-        </section>
-        <section className={styles.statistics_counters}>
-          <p className="text text_type_main-medium">Выполнено за все время:</p>
-          <p
-            className={styles.digits_glow + " " + "text text_type_digits-large"}
-          >
-            123456
-          </p>
-        </section>
-        <section className={styles.statistics_counters}>
-          <p className="text text_type_main-medium">Выполнено за сегодня:</p>
-          <p
-            className={styles.digits_glow + " " + "text text_type_digits-large"}
-          >
-            123456
-          </p>
-        </section>
       </article>
     </>
   );
