+++ src/components/app/pages/profile-form/profile-form.module.css
@@ -1,0 +1,18 @@
+
+.profile_form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 24px;
+  composes: mt-30 from global;
+  width: 480px;
+}
+
+.button_form {
+  opacity: 0;
+  transition: opacity 1s ease-in-out;
+}
+
+.button_form_show {
+  opacity: 1;
+}
