+++ src/components/app/pages/profile/profile.module.css
@@ -1,30 +1,18 @@
 .profile {
   display: flex;
-  margin: 120px 0 0 0;
   gap: 60px;
+  width: 1232px;
 }
 
 .profile_navlinks {
   width: 320px;
+  composes: mt-30 from global;
 }
 
 .profile_inputs_wrapper {
   display: flex;
   align-items: center;
   justify-content: center;
-  width: 480px;
-}
-
-.loading_gif {
-  width: 100px;
-  height: 100px;
-}
-
-.profile_form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 24px;
 }
 
 .link {
@@ -36,13 +24,4 @@
 
 .link_disabled {
   color: grey;
-}
-
-.button_form {
-  opacity: 0;
-  transition: opacity 1s ease-in-out;
-}
-
-.button_form_show {
-  opacity: 1;
 }
