+++ src/components/order-current/order-current.tsx
@@ -1,0 +1,3 @@
+export default function OrderCurrent() {
+  return <></>;
+}
